Making simple bar graph dynamic
