+++ src/components/BlockExplorer.tsx
@@ -1,49 +1,11 @@
 import React from 'react';
 import './BlockExplorer.css';
-import SimpleBarGraph from './Graphs/SimpleBarGraph';
 import PlainGraphTitle from './GraphTitles/PlainGraphTitle';
 import PolygonGraph from './Graphs/PolygonGraph';
 import Hero from './Hero';
 import HeroGraph from './Graphs/HeroGraph';
 import LatestBlocks from './LatestBlocks';
-const barGraphData = [
-    2,
-    4,
-    5,
-    6,
-    7,
-    8,
-    9,
-    11,
-    14,
-    15,
-    16,
-    20,
-    20,
-    22,
-    23,
-    25,
-    27,
-    29,
-    30,
-    31,
-    32,
-    33,
-    33,
-    33,
-    34,
-    35,
-    36,
-    37,
-    38,
-    39,
-    40,
-    42,
-    44,
-    44,
-    46,
-    47
-];
+import CirculatingTariGraph from "./CirculatingTariGraph";
 
 const hashRateData = [
     50,
@@ -99,7 +61,7 @@ export default function BlockExplorer() {
             <div className="twoCol">
                 <PolygonGraph width={500} height={220} yAxisTicks={6} data={hashRateData} />
 
-                <SimpleBarGraph width={500} height={220} yAxisTicks={6} data={barGraphData} />
+                <CirculatingTariGraph />
             </div>
         </div>
     );
