+++ src/components/Graphs/SimpleBarGraph.tsx
@@ -1,45 +1,24 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import './SimpleBarGraph.css';
 import { scaleLinear } from 'd3-scale';
 import PlainGraphTitle from '../GraphTitles/PlainGraphTitle';
-import { fetchTokensInCirculation } from '../../helpers/api';
 import numeral from 'numeral';
 import config from '../../config';
 const { tokenName } = config;
 
 interface Props {
-    data: number[];
+    data: any[];
     width: number;
     height: number;
     yAxisTicks: number;
 }
 export default function SimpleBarGraph({ width, height, data, yAxisTicks }: Props) {
-    const [totalTokens, setTotalTokens] = useState(([] as unknown) as any);
-
-    const loadCirculationData = useCallback(async () => {
-        const tokenData = await fetchTokensInCirculation();
-        const heightsArr: any[] = [];
-        const totalsArr: number[] = [];
-
-        tokenData.map((token) => {
-            const { height, tokensInCirculation } = token;
-            heightsArr.push(height);
-            totalsArr.push(tokensInCirculation);
-        });
-
-        setTotalTokens(totalsArr);
-    }, []);
-
-    useEffect(() => {
-        loadCirculationData().then((r) => {});
-    }, [loadCirculationData]);
-
     const yScale = scaleLinear()
-        .domain([0, Math.max(...totalTokens)])
+        .domain([0, Math.max(...data)])
         .range([height, 0]);
-    const barWidth = Math.floor(width / totalTokens.length);
+    const barWidth = Math.floor(width / data.length);
 
-    const highestNum = Math.max(...totalTokens);
+    const highestNum = Math.max(...data);
 
     function round5({ num }: { num: any }) {
         return Math.ceil(num / 5) * 5;
@@ -81,7 +60,7 @@ export default function SimpleBarGraph({ width, height, data, yAxisTicks }: Prop
         }
         return nums;
     }
-    // eslint-disable-next-line no-undef
+
     const title = `Circulating ${tokenName}`;
     return (
         <div className="graphWrapper">
@@ -93,7 +72,7 @@ export default function SimpleBarGraph({ width, height, data, yAxisTicks }: Prop
             <svg className="circulateSimpleBars" height={height} width={width}>
                 {renderYAxis()}
 
-                {totalTokens.map((total, i) => {
+                {data.map((total, i) => {
                     return (
                         <g key={i} className="barHolder">
                             <g
